requirements.txt
tensorflow 
opencv-python 
matplotlib
numpy
pillow
